GiftTracker/pyproject.toml
[tool.poetry]
name = "repl-nix-gifttracker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.38.0"
psycopg2-binary = "^2.9.9"
plotly = "^5.24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
